ocr-service/requirements.txt
Django==5.2.5
djangorestframework==3.16.1
django-cors-headers==4.7.0
channels==4.3.1
channels-redis==4.3.0
daphne==4.2.1
redis==6.4.0
python-dotenv==1.1.1
PyJWT==2.10.1
Pillow==11.3.0
pytesseract==0.3.13
pdf2image==1.17.0
transformers==4.54.1
torch==2.8.0
torchvision==0.23.0
opencv-python-headless==4.12.0.88
numpy==2.2.3
requests==2.32.4
gunicorn==23.0.0
whitenoise==6.9.0